Don't add input parameter argument if not necessary
If we don't need the input parameter argument don't add it to the
interface. It makes both the client and server implementations of these
methods a bit simpler, though methods with no arguments are fairly
unlikely from a client perspective as they usually come up in health
checks. It makes the codegen logic a bit trickier.

Not sure if it's worth it.