Stop cancelling requests before we're done with them
Before this change the call to `defer.Cancel()` in the timeout doer
would cause us to cancel the underlying http request before we were
actually done processing it. For large response bodies this caused us to
get "request cancelled" errors while parsing the output.

We still want to call cancel based on the go documentation
https://github.com/golang/net/blob/master/context/go17.go#L53-L54 to
avoid temporarily leaking resources, so I decided to move the
cancellation up to the top level of code that handles http objects. This
way we only call cancel once we're actually done with the http request.