Add field AllowSecondaryIndexScan to x-db configs.
(Generated files in subsequent commit)

We already have an opt-in field for scans over the main index. This
commit adds support for scanning over a secondary index using a field
AllowSecondaryIndexScan, which is a list of index names to generate
scan code for.

The use case for this that I have in mind is sparse indexes, so that
the index will have fewer items than the whole table. In that case, I
want to have code to scan over the much smaller GSI rather than the
entire table.