Add support for x-nullable fields.
The only change is to the function that determines if a field is a
pointer or not in `server/gendb/templatefuncs.go`. The new test model
in `samples/db.yml` includes a nullable attribute as a range key for
two indexes: one with a non-composite hash key, and one with composite.