Add a replace for go-openapi/wag
This is a breaking change which reverts unintentional breaking changes that were introduced in
5.0.0. Namely, 5.0.0 changed the version of `github.com/go-openapi/swag` package in a way that
changed some minor aspects of the generated code (e.g. capitalization). We are regarding that change
as overly breaking for now, and so we're rolling it back so that users of `wag` can easily upgrade
to using `wag` with modules without worrying about it breaking all their stuff.