Don't call (*Test).Fatal from inside a goroutine
This is a new vet check in go1.16 - these calls weren't doing what we hoped they were
doing (i.e. exiting the whole test). There's no good way, it seems, to exit the whole test from
inside a goroutine. So instead, just mark the test as failed and exit the goroutine.

Also, we don't use the asset file metadata from go-bindata, but it causes a bunch of extraneous
diffs when regenerating, so stop storing that.