Remove references to Glide.
Glide is/was a package manager for Go way back when. Wag could integrate with the files to make sure
that the project which had the generated code in it had dependencies that Wag expected. Wag no
longer offers this functionality and Glide is not supported by Wag any more.