Quote certain field names in generated index.d.ts files.
Our previous output failed in the case that a field name was not a valid JavaScript identifier, for
example if it contained a dash ("-").

In this fix, I attempt to leave most things unquoted, but quote when necessary. However, exactly
what needs to be quoted is a bit complicated since it involves a specific definition of Unicode
"letter", so this may quote slightly more often to be on the cautious side while having a simple implementation.