remove default table names; validate instead
The default table name provided is not the standard table naming convention used at Clever. Because of the default behavior, this lead to cryptic IAM errors at run time when an engineer forgot to specify the table name when adding a new table. These errors have a table name that was often nearly identical aside from capitalization to the real table name. Getting rid of the default has no real effect aside from making this step easier to perform and debug.