chore: Don't generate JavaScript clients for subrouters
Since the strategy for subrouter JavaScript clients (at least, so far), has been
to include the subrouters in the root client generation — which is the opposite
of how the Go clients are generated — we don't need to generate the subrouter
clients. This change autosets conf.generateJSClient to false if conf.subrouter
is true.